+++ changelog.md
@@ -1,11 +1,17 @@
+## 1.0.2
+24/12/20
+
+- Added `to_ffi_str()` and `to_ffi_string()` to &str
+- Added `to_ffi_string()` and `into_ffi_string()` to String
+
+<br>
+
 ## 1.0.1
 24/12/20
 
 - Added `Debug` to both types
 - Added `Clone` to both types
 - Added `Copy` to FFIStr
-
-- 
 
 <br>